ArtosBackend/requirements.txt
Flask>=2.3
flask-cors>=4.0.0

# LangChain ecosystem
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-google-genai>=2.0.0

# Vector store
faiss-cpu>=1.7.4

# Parsing & text utils
pymupdf>=1.23.0
python-docx>=1.1.0
tiktoken>=0.5.0

